refac: 상관관계 중복 코드 추출
